// components/trigger/enum.tsx
export enum Action {
  click = 'click',
  focus = 'focus',
  hover = 'hover',
  contextMenu = 'contextMenu',
}

export enum ShowAction {
  click = 'click',
  focus = 'focus',
  mouseEnter = 'mouseEnter',
  contextMenu = 'contextMenu',
}

export enum HideAction {
  click = 'click',
  blur = 'blur',
  mouseLeave = 'mouseLeave',
}